requirements.txt
google-auth
google-cloud
google-cloud-vision
opencv-python